planner/requirements.txt
fastapi
uvicorn[standard]
google-cloud-pubsub
google-cloud-firestore
python-dotenv
requests
